+++ Wx.Exercises/Wx.Exercises.Services/Proxies/WxApiProxy/Models/Product.cs
@@ -6,8 +6,8 @@ namespace Wx.Exercises.Services.Proxies.WxProxy.Models
     public class Product
     {
         public string Name { get; init; }
-        public decimal Price { get; init; }
-        public decimal Quantity { get; init; }
+        public double Price { get; init; }
+        public double Quantity { get; init; }
     }
 
 }
